# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AttachmentUploadItem": {
        "properties": {
          "filename": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "minimum": 1,
            "type": "integer"
          },
          "size_bytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "filename",
          "size_bytes"
        ],
        "type": "object"
      },
      "AttachmentUploadRequest": {
        "additionalProperties": false,
        "properties": {
          "file": {
            "items": {},
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "file"
        ],
        "type": "object"
      },
      "AttachmentUploadResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AttachmentUploadItem"
            },
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "maximum": 599,
            "minimum": 100,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "format": "uri-reference",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "in": "cookie",
        "name": "racebin_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Racebin contributors"
    },
    "description": "API-first, self-hosted paste bin and file share.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Racebin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/pastes/{paste_id}/archive": {
      "get": {
        "operationId": "get_archive",
        "parameters": [
          {
            "description": "Paste ID",
            "in": "path",
            "name": "paste_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Short-lived token returned by the final permitted read.",
            "in": "query",
            "name": "read_token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/zip": {}
            },
            "description": "ZIP containing paste text and attachments",
            "headers": {
              "Content-Disposition": {
                "description": "Archive filename",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid bearer credential"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Paste not found, not visible, or grant invalid"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Archive input exceeds 64 MiB"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Archive could not be generated"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "attachments"
        ],
        "x-racebin-scopes": [
          "paste:read"
        ]
      }
    },
    "/pastes/{paste_id}/attachments": {
      "post": {
        "operationId": "upload_attachments",
        "parameters": [
          {
            "description": "Paste ID",
            "in": "path",
            "name": "paste_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Current ETag or *",
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Required for session-cookie mutations",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/AttachmentUploadRequest"
              }
            }
          },
          "description": "One or more repeated file parts",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentUploadResponse"
                }
              }
            },
            "description": "Attachments uploaded",
            "headers": {
              "ETag": {
                "description": "New paste entity tag",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Malformed or empty upload"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Uploads disabled or insufficient permission"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Paste not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Attachment conflict"
          },
          "412": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "ETag does not match"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Attachment limit exceeded"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid multipart field"
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "If-Match is required"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "attachments"
        ],
        "x-racebin-scopes": [
          "paste:write"
        ]
      }
    },
...
    "/pastes/{paste_id}/qr": {
      "get": {
        "operationId": "get_qr",
        "parameters": [
          {
            "description": "Paste ID",
            "in": "path",
            "name": "paste_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/png": {}
            },
            "description": "PNG QR code for the public paste URL"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid bearer credential"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Paste not found, not visible, or QR disabled"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "QR generation failed"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "attachments"
        ],
        "x-racebin-scopes": [
          "paste:read"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}